+++ CHANGELOG.md
@@ -3,7 +3,7 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 #### 1.9.11-SNAPSHOT
 ``
 - **[UPDATE]** Updated to Kotlin 1.4.10.
-- **[UPDATE]** Updated to Kotlin Coroutines 1.3.9.
+- **[UPDATE]** Updated to Kotlin Coroutines 1.4.0.
 - **[UPDATE]** Updated to Gradle 6.7.
 - **[MISC]** Groovy Gradle scripts and configuration files migrated to Kotlin.
 - **[FEATURE]** (`ktx-collections`) Added `GdxIdentityMap` and `GdxArrayMap` aliases for LibGDX `IdentityMap` and `ArrayMap` collections.